+++ packages/cli/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Hashing will now ignore files that matched a pattern found in the root `.gitignore`.
+- Updated the JSON schemas to use the new package manager versions.
+
 #### 🐞 Fixes
 
-- Fixed an issue with the `.moon/workspace.yml` template during `moon init` generating with invalid
-  whitespace.
-- Updated the JSON schemas to use the new package manager versions.
+- Fixed an issue with the `.moon/workspace.yml` template being generating with invalid whitespace
+  during `moon init`.